S302T700-94 #time 10m Refactored Inventory to have Integer ID. ?re-engineering

+++ backend/src/main/java/com/seng302/wasteless/dto/PostListingsDto.java
@@ -22,7 +22,7 @@ public class PostListingsDto {
    
 
     @NotBlank(message = "Inventory item ID is Mandatory")
-    private Long inventoryItemId;
+    private Integer inventoryItemId;
 
     @Positive
     @Max(1000000000)            //max = inventoryItem.quantity()
