+++ backend/src/main/java/com/seng302/wasteless/service/InventoryService.java
@@ -26,7 +26,7 @@ public class InventoryService {
      * @param id        The id of the Inventory item to be found
      * @return          The found Inventory item, if any, otherwise null
      */
-    public Inventory findInventoryById(Long id) {
+    public Inventory findInventoryById(Integer id) {
         return inventoryRepository.findFirstById(id);
     }
